04/08 更改position embedding

+++ vae.py
@@ -240,8 +240,8 @@ class PositionalEncoding(nn.Module):
         position = torch.arange(0, max_len, dtype=torch.float).unsqueeze(1)
         dimention=torch.arange(0, d_model)
 
-        pe[:, 0::2] = torch.sin(position / torch.pow(10000, 2*(dimention//2)[::2]/d_model))
-        pe[:, 1::2] = torch.cos(position / torch.pow(10000, 2*(dimention//2)[1::2]/d_model))
+        pe[:, 0::2] = torch.sin(position / torch.pow(10000, 2*(dimention//2)[::2].float()/d_model))
+        pe[:, 1::2] = torch.cos(position / torch.pow(10000, 2*(dimention//2)[1::2].float()/d_model))
         pe = pe.unsqueeze(0)
         self.register_buffer('pe', pe)
